requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.1
requests~=2.32.3
schedule~=1.2.2
aiogram~=3.11.0
python-cron
DateTime~=5.5
APScheduler~=3.10.4